requirements.txt
flask==2.3.3
gunicorn
google-generativeai
requests
cloudscraper